WIP: makes needleslot left-aligned in needleunit
Prior to this change, needleunits have centred needleslots.
This leads to thin walls on each side of the needlebed, which are
prone to breaking off. This is at its worst on 4.5mm gauge.

This commit adjusts the needleslot to be left-aligned.
This changes the overall geometry of the needlebed and required
additional adjustments to the spongebar and backcover parts.
A small change may still be needed on the left mounting bracket.
As this part is currently undergoing refactoring, this was deferred.
For this reason, the new parts have not yet been test-printed!

The needleslot not being left-aligned has one major downside:
The left-most needle on each needlebed has no support when the
knitting machine is disassembled. It will fall out, unless secured
by tape or other means.